n2v conf wip
